requirements.txt
pytest>=7.0.0
requests>=2.28.0
PyYAML>=6.0
allure-pytest>=2.13.0
pytest-html>=3.2.0
